requirements.txt
boto==2.49.0
certifi==2020.6.20
chardet==3.0.4
Django==3.2.5
django-crispy-forms==1.12.0
docutils==0.16
gunicorn==20.1.0
idna==2.10
Pillow==8.0.1
python-dateutil==2.8.1
pytz==2020.1
requests==2.24.0
six==1.15.0
urllib3==1.25.11
whitenoise==5.3.0
psycopg2-binary==2.9.1
djangorestframework==3.12.4